requirements.txt
streamlit
spacy
py2neo
python-decouple
spacy-transformers
sentence-transformers
streamlit-option-menu